karting classes added
