+++ MFC_test/Kart.cpp
@@ -1,0 +1,1 @@
+#include"Kart.h"
