+++ MFC_test/Kart.h
@@ -1,0 +1,11 @@
+#pragma once
+class Kart {
+public:
+	Vec3 Pos;
+	Vec3 dir;
+	float currSpeed;
+	float maxSpeed;
+	WayPoint * startWP;
+
+	Kart(Vec3 Pos);
+};
