+++ MFC_test/MFC_test.vcxproj
@@ -186,15 +186,19 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="DlgName.h" />
+    <ClInclude Include="Kart.h" />
     <ClInclude Include="KartDlg.h" />
     <ClInclude Include="MFC_test.h" />
     <ClInclude Include="MFC_testDlg.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
+    <ClInclude Include="WayPoint.h" />
+    <ClInclude Include="World.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="DlgName.cpp" />
+    <ClCompile Include="Kart.cpp" />
     <ClCompile Include="KartDlg.cpp" />
     <ClCompile Include="MFC_test.cpp" />
     <ClCompile Include="MFC_testDlg.cpp" />
@@ -204,6 +208,8 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="WayPoint.cpp" />
+    <ClCompile Include="World.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="MFC_test.rc" />
