+++ MFC_test/MFC_test.vcxproj.filters
@@ -13,6 +13,15 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
+    <Filter Include="KartClasses">
+      <UniqueIdentifier>{044f13bf-051c-4cee-9c5c-e8764826f996}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="KartClasses\src">
+      <UniqueIdentifier>{5b8b0afe-a4a5-4c69-ac55-c71e1649a0e4}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="KartClasses\incl">
+      <UniqueIdentifier>{0075816b-f087-48a0-9600-f6f50771222b}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <Text Include="ReadMe.txt" />
@@ -39,6 +48,15 @@
     <ClInclude Include="KartDlg.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Kart.h">
+      <Filter>KartClasses\incl</Filter>
+    </ClInclude>
+    <ClInclude Include="WayPoint.h">
+      <Filter>KartClasses\incl</Filter>
+    </ClInclude>
+    <ClInclude Include="World.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="MFC_test.cpp">
@@ -54,6 +72,15 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="KartDlg.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Kart.cpp">
+      <Filter>KartClasses\src</Filter>
+    </ClCompile>
+    <ClCompile Include="WayPoint.cpp">
+      <Filter>KartClasses\src</Filter>
+    </ClCompile>
+    <ClCompile Include="World.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
