+++ MFC_test/WayPoint.cpp
@@ -1,0 +1,12 @@
+#include "stdafx.h"
+#include "WayPoint.h"
+
+
+WayPoint::WayPoint()
+{
+}
+
+
+WayPoint::~WayPoint()
+{
+}
