+++ MFC_test/WayPoint.h
@@ -1,0 +1,12 @@
+#pragma once
+class WayPoint : public CDialog {
+public:
+	Vec3 Pos;
+	bool is_StartLine;
+	WayPoint * nextWP;
+	WayPoint * prevWP;
+
+	WayPoint(Vec3 Pos);
+
+};
+
