+++ MFC_test/World.cpp
@@ -1,0 +1,12 @@
+#include "stdafx.h"
+#include "World.h"
+
+
+World::World()
+{
+}
+
+
+World::~World()
+{
+}
